# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Global College LMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /assignments/submissions:
    get:
      operationId: listSubmissions
      tags: [assignments]
      summary: List assignment submissions (teacher/admin)
      parameters:
        - name: assignmentId
          in: query
          schema:
            type: integer
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssignmentSubmission"
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentBody"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    put:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Delete assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assignments/{id}/submit:
    post:
      operationId: submitAssignment
      tags: [assignments]
      summary: Submit assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAssignmentBody"
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentSubmission"
  /assignments/{id}/grade:
    post:
      operationId: gradeAssignment
      tags: [assignments]
      summary: Grade assignment submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeAssignmentBody"
      responses:
        "200":
          description: Graded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignmentSubmission"
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches/campuses
      responses:
        "200":
          description: Branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create branch (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    put:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBranchBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        phone:
          type: string
        cnic:
          type: string
        branchId:
          type: integer
      required: [name, email, password, role]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, teacher, admin]
        phone:
          type: string
        cnic:
          type: string
        avatar:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, name, email, role, isActive, createdAt]
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        totalMarks:
          type: integer
        createdAt:
          type: string
      required: [id, courseId, title, totalMarks, createdAt]
    CreateAssignmentBody:
      type: object
      properties:
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        dueDate:
          type: string
        totalMarks:
          type: integer
      required: [courseId, title, totalMarks]
    AssignmentSubmission:
      type: object
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        userId:
          type: integer
        fileUrl:
          type: string
        notes:
          type: string
        submittedAt:
          type: string
        marks:
          type: integer
        feedback:
          type: string
        status:
          type: string
          enum: [submitted, graded, late]
        userName:
          type: string
        assignmentTitle:
          type: string
      required: [id, assignmentId, userId, status]
    SubmitAssignmentBody:
      type: object
      properties:
        fileUrl:
          type: string
        notes:
          type: string
      required: [fileUrl]
    GradeAssignmentBody:
      type: object
      properties:
        submissionId:
          type: integer
        marks:
          type: integer
        feedback:
          type: string
      required: [submissionId, marks]
...
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        mapUrl:
          type: string
        isMain:
          type: boolean
        createdAt:
          type: string
      required: [id, name, city, address, createdAt]
    CreateBranchBody:
      type: object
      properties:
        name:
          type: string
        city:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        mapUrl:
          type: string
        isMain:
          type: boolean
      required: [name, city, address]
...